+++ caravel/models.py
@@ -582,7 +582,6 @@ class Database(Model, AuditMixinNullable):
     """An ORM object that stores Database related information"""
 
     __tablename__ = 'dbs'
-
     id = Column(Integer, primary_key=True)
     database_name = Column(String(250), unique=True)
     sqlalchemy_uri = Column(String(1024))
@@ -807,8 +806,7 @@ class SqlaTable(Model, Queryable, AuditMixinNullable, ImportMixin):
 
     @property
     def full_name(self):
-        return utils.get_datasource_full_name(
-            self.database, self.table_name, schema=self.schema)
+        return "[{obj.database}].[{obj.table_name}]".format(obj=self)
 
     @property
     def dttm_cols(self):
@@ -1374,7 +1372,6 @@ class DruidCluster(Model, AuditMixinNullable):
     """ORM object referencing the Druid clusters"""
 
     __tablename__ = 'clusters'
-
     id = Column(Integer, primary_key=True)
     cluster_name = Column(String(250), unique=True)
     coordinator_host = Column(String(255))
@@ -1487,8 +1484,9 @@ class DruidDatasource(Model, AuditMixinNullable, Queryable):
 
     @property
     def full_name(self):
-        return utils.get_datasource_full_name(
-            self.cluster_name, self.datasource_name)
+        return (
+            "[{obj.cluster_name}]."
+            "[{obj.datasource_name}]").format(obj=self)
 
     @property
     def time_column_grains(self):
