+++ tests/base_tests.py
@@ -20,8 +20,8 @@ cli = imp.load_source('cli', BASE_DIR + "/bin/caravel")
 
 
 class CaravelTestCase(unittest.TestCase):
-    requires_examples = True
-    examples_loaded = True
+    requires_examples = False
+    examples_loaded = False
 
     def __init__(self, *args, **kwargs):
         if (
@@ -118,15 +118,6 @@ class CaravelTestCase(unittest.TestCase):
         )
         session.expunge_all()
         return slc
-
-    def get_table_by_name(self, name):
-        return db.session.query(models.SqlaTable).filter_by(
-            table_name=name).first()
-
-    def get_druid_ds_by_name(self, name):
-        return db.session.query(models.DruidDatasource).filter_by(
-            datasource_name=name).first()
-
 
     def get_resp(self, url):
         """Shortcut to get the parsed results while following redirects"""
